Created a semi complete Vault and Account class. Ran into several syntax errors that I just am not seeing.
Vault has the ability to be instantiated, open a JSON file and parse the
JSON file in an assumed fixed format and create an arraylist of Account
class objects. It also has the ability to add and remove objects from
the arraylist. It also has a method to return the arraylist of accounts
back to main.

Accounts has the ability to be instantiated and has a method that aids
the ArrayList.remove method.

It does not have the ability to edit Accounts in the ArrayList and does
not perform authentication of any kind. Furthermore, it does not output
the accounts to a JSON file.